+++ include/libzm2midi.h
@@ -41,6 +41,8 @@ typedef uint8_t Zm2M1MidiCh[8];
 typedef uint8_t Zm2M1TrackParam[40];
 typedef uint8_t Zm2M1EffectParam[25];
 
+enum { ZM2_USE_PREVIOUS_DEVID = -1 };
+
 static inline void  //
 zm2_midi_rec(void) {
   zm2__func(0x16);
@@ -58,9 +60,9 @@ zm2_midi_trns(uint32_t size, const uint8_t* data) {
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_send_rd_exc(uint32_t size, uint16_t devid, uint16_t modelid,
+zm2_send_rd_exc(uint32_t size, int8_t devid, uint8_t modelid,
                 const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x1c, size, (devid << 16) + modelid, data);
+  return zm2__func_d2d3a1_wa(0x1c, size, ((uint8_t)devid << 8) | modelid, data);
 }
 
 static inline zm2__nonnull int32_t  //
@@ -69,124 +71,122 @@ zm2_send_exc(uint32_t size, const uint8_t* data) {
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_p_rsv(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x1e, size, devid, data);
+zm2_sc55_p_rsv(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x1e, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_reverb(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x1f, size, devid, data);
+zm2_sc55_reverb(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x1f, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_chorus(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x20, size, devid, data);
+zm2_sc55_chorus(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x20, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_part_setup(uint32_t size, uint16_t part, uint16_t devid,
+zm2_sc55_part_setup(uint32_t size, uint8_t part, int8_t devid,
                     const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x21, size, (part << 16) + devid, data);
+  return zm2__func_d2d3a1_wa(0x21, size, (part << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_drum_setup(uint32_t size, uint8_t map, uint8_t note, uint16_t devid,
+zm2_sc55_drum_setup(uint32_t size, uint8_t map, uint8_t note, int8_t devid,
                     const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x22, size, (map << 24) + (note << 16) + devid,
-                             data);
+  uint32_t d3 = (((map << 8) | note) << 16) | (uint8_t)devid;
+  return zm2__func_d2d3a1_wa(0x22, size, d3, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_print(uint32_t size, uint8_t devid, const char* message) {
-  return zm2__func_d2d3a1_wa(0x23, size, devid, message);
+zm2_sc55_print(uint32_t size, int8_t devid, const char* message) {
+  return zm2__func_d2d3a1_wa(0x23, size, (uint8_t)devid, message);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_sc55_display(uint8_t devid, const Zm2Sc55Display* pattern) {
-  return zm2__func_d3a1(0x24, devid, pattern);
+zm2_sc55_display(int8_t devid, const Zm2Sc55Display* pattern) {
+  return zm2__func_d3a1(0x24, (uint8_t)devid, pattern);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_p_rsv(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x25, size, devid, data);
+zm2_mt32_p_rsv(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x25, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_reverb(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x26, size, devid, data);
+zm2_mt32_reverb(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x26, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_setup(uint32_t size, uint8_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x27, size, devid, data);
+zm2_mt32_setup(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x27, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_drum(uint32_t size, uint16_t note, uint16_t devid,
-              const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x28, size, (note << 16) + devid, data);
+zm2_mt32_drum(uint32_t size, uint8_t note, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x28, size, (note << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_common(uint32_t size, uint16_t prog, uint16_t devid,
+zm2_mt32_common(uint32_t size, uint8_t prog, int8_t devid,
                 const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x29, size, (prog << 16) + devid, data);
+  return zm2__func_d2d3a1_wa(0x29, size, (prog << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_partial(uint32_t size, uint8_t prog, uint8_t part, uint16_t devid,
+zm2_mt32_partial(uint32_t size, uint8_t prog, uint8_t part, int8_t devid,
                  const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x2a, size, (prog << 24) + (part << 16) + devid,
-                             data);
+  uint32_t d3 = (((prog << 8) | part) << 16) | (uint8_t)devid;
+  return zm2__func_d2d3a1_wa(0x2a, size, d3, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_patch(uint32_t size, uint16_t patch, uint16_t devid,
+zm2_mt32_patch(uint32_t size, uint8_t patch, int8_t devid,
                const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x2b, size, (patch << 16) + devid, data);
+  return zm2__func_d2d3a1_wa(0x2b, size, (patch << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_mt32_print(uint32_t size, uint8_t devid, const char* message) {
-  return zm2__func_d2d3a1_wa(0x2c, size, devid, message);
+zm2_mt32_print(uint32_t size, int8_t devid, const char* message) {
+  return zm2__func_d2d3a1_wa(0x2c, size, (uint8_t)devid, message);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_setup(uint32_t size, uint16_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x2d, size, devid, data);
+zm2_u220_setup(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x2d, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_common(uint32_t size, uint16_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x2e, size, devid, data);
+zm2_u220_common(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x2e, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_d_setup(uint32_t size, uint16_t devid, const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x2f, size, devid, data);
+zm2_u220_d_setup(uint32_t size, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x2f, size, (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_p_setup(uint32_t size, uint16_t part, uint16_t devid,
+zm2_u220_p_setup(uint32_t size, uint8_t part, int8_t devid,
                  const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x30, size, (part << 16) + devid, data);
+  return zm2__func_d2d3a1_wa(0x30, size, (part << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_print(uint32_t size, uint8_t devid, const char* message) {
-  return zm2__func_d2d3a1_wa(0x31, size, devid, message);
+zm2_u220_print(uint32_t size, int8_t devid, const char* message) {
+  return zm2__func_d2d3a1_wa(0x31, size, (uint8_t)devid, message);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_timbre(uint32_t size, uint16_t prog, uint16_t devid,
+zm2_u220_timbre(uint32_t size, uint8_t prog, int8_t devid,
                 const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x32, size, (prog << 16) + devid, data);
+  return zm2__func_d2d3a1_wa(0x32, size, (prog << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
-zm2_u220_drum(uint32_t size, uint16_t note, uint16_t devid,
-              const uint8_t* data) {
-  return zm2__func_d2d3a1_wa(0x33, size, (note << 16) + devid, data);
+zm2_u220_drum(uint32_t size, uint8_t note, int8_t devid, const uint8_t* data) {
+  return zm2__func_d2d3a1_wa(0x33, size, (note << 16) | (uint8_t)devid, data);
 }
 
 static inline zm2__nonnull int32_t  //
@@ -195,8 +195,8 @@ zm2_m1_midi_ch(const Zm2M1MidiCh* data) {
 }
 
 static inline int32_t  //
-zm2_send_to_m1(uint8_t devid) {
-  return zm2__func_d3(0x35, devid);
+zm2_send_to_m1(int8_t devid) {
+  return zm2__func_d3(0x35, (uint8_t)devid);
 }
 
 static inline zm2__nonnull int32_t  //
@@ -221,12 +221,12 @@ zm2_eox_wait(uint16_t wait) {
 
 static inline int32_t  //
 zm2_sc55_init(int8_t devid) {
-  return zm2__func_d3(0x51, devid);
+  return zm2__func_d3(0x51, (uint8_t)devid);
 }
 
 static inline int32_t  //
 zm2_mt32_init(int8_t devid) {
-  return zm2__func_d3(0x52, devid);
+  return zm2__func_d3(0x52, (uint8_t)devid);
 }
 
 static inline int32_t  //